+++ lib/components/card_view/enlarge_controller_edit.dart
