+++ lib/components/card_view/event/event_information_enlarge.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:grouping_project/components/card_view/enlarge_fragment_body.dart';
+import 'package:grouping_project/components/card_view/enlarge_controller_view.dart';
 import 'package:grouping_project/components/card_view/enlarge_context_template.dart';
 import 'package:grouping_project/model/model_lib.dart';
-import 'package:grouping_project/components/card_view/event_edit.dart';
+import 'package:grouping_project/components/card_view/event_information.dart';
 
 class EventInformationEnlarge extends StatefulWidget {
   /// (未完成!!)
