+++ lib/components/card_view/event_information.dart
@@ -1,2 +1,4 @@
-export 'package:grouping_project/components/card_view/event_information_shrink.dart';
-export 'package:grouping_project/components/card_view/event_information_enlarge.dart';
+export 'package:grouping_project/components/card_view/event/event_information_shrink.dart';
+export 'package:grouping_project/components/card_view/event/event_information_enlarge.dart';
+export 'package:grouping_project/components/card_view/enlarge_context_template.dart';
+export 'package:grouping_project/components/card_view/event/event_edit.dart';
