+++ lib/components/card_view/mission/mission_information.dart
