+++ lib/pages/home/personal_dashboard/home_mission_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:grouping_project/components/card_view/card_view_template.dart';
 import 'package:grouping_project/model/model_lib.dart';
-import 'package:grouping_project/components/card_view/mission_information.dart';
+import 'package:grouping_project/components/card_view/mission/mission_information.dart';
 
 class MissionPage extends StatefulWidget {
   const MissionPage({super.key});
